use ScriptEngineFactory and ScriptEngine to load python udfs

+++ udf/age/age.py
@@ -1,12 +1,11 @@
 #!/usr/bin/env python
-# -*- coding: utf-8 -*-
 
 from io.crate.udf import UserDefinedScalarFunction, UserDefinedAggregationFunction
 from io.crate.operation.aggregation import AggregationState
 from io.crate import DataType
 from io.crate.metadata import FunctionIdent, FunctionInfo
 from java.util import Arrays
-
+from java.lang import Long
 
 
 class PythonMax(UserDefinedAggregationFunction):
@@ -36,7 +35,7 @@ class PythonMax(UserDefinedAggregationFunction):
         return 'python_max'
 
     def ident(self):
-        return FunctionIdent(self.name(), Arrays.asList(DataType.LONG))
+        return FunctionIdent(self.name(), Arrays.asList(DataType.INTEGER))
 
     def info(self):
         return FunctionInfo(self.ident(), DataType.LONG, True)
@@ -73,4 +72,4 @@ class MathMin(UserDefinedScalarFunction):
         return None
 
     def evaluate(self, args):
-        return min(args[0].value(), args[1].value())
+        return Long(min(args[0].value(), args[1].value()))
